INT-1306: don't convert similar types

+++ evaluator/test_setup.go
@@ -249,7 +249,7 @@ schema:
      -
         Name: a
         MongoType: int
-        SqlType: int
+        SqlType: numeric
      -
         Name: b
         MongoType: int
